feat: 用户登陆

+++ api/src/db/user.rs
@@ -1,6 +1,5 @@
-use chrono::{DateTime, NaiveDateTime, Utc};
+use chrono::{DateTime, Utc};
 use sqlx::FromRow;
-use toml::value::Datetime;
 
 use crate::db::fetch_db_pool;
 
@@ -34,7 +33,7 @@ pub enum UserStatus {
 
 impl User{
 
-    pub async fn get_by_account_id(account_id: i32)->anyhow::Result<Self> {
+    pub async fn get_by_account_id(account_id: i64)->anyhow::Result<Self> {
 
         let user = sqlx::query_as("SELECT id, account_id, credential, salt, nickname, avatar, signature, status, info, create_at, update_at, delete_at FROM api.user \
         WHERE account_id = $1")
